extract generation code to notebook file
