+++ app/src/generate.py
@@ -45,25 +45,6 @@ If the question would be easier to answer with additional policy or program cont
 Otherwise, set needs_context to false.
 """
 
-GENERATE_QUESTION_ANSWER_PROMPT = """
-Using the provided chunk, generate at least 5 unique questions and answers, avoid rephrasing or changing the punctuation of the question to ensure distinct questions and answers.
-Respond with a list of JSON dictionaries in the following format  (do not wrap in JSON markers):
-
-question: The generated question based on the content.
-answer: The answer to the question, derived from the content.
-document_name: The name of the document.
-document_source: The source or URL of the document.
-
-Example
-question: What's a base period for SDI?
-answer: "- A base period covers 12 months and is divided into four quarters.
-- The base period includes wages subject to SDI tax that were paid about 5 to 18 months before the client's disability claim began
-- For a DI claim to be valid, they must have at least $300 in wages in the base period.
-- Benefit amounts are based on the quarter with their highest wages earned within their base period."
-document_name: Disability Insurance Benefit Payment Amounts
-document_source: https://edd.ca.gov/en/disability/Calculating_DI_Benefit_Payment_Amounts/
-"""
-
 
 def get_models() -> dict[str, str]:
     """
@@ -132,45 +113,6 @@ def completion_args(llm: str) -> dict[str, Any]:
     if llm.startswith("ollama/"):
         return {"api_base": os.environ["OLLAMA_HOST"]}
     return {}
-
-
-class QuestionAnswerAttributes(BaseModel):
-    question: str
-    answer: str
-    document_name: str
-    document_source: str
-
-
-class QuestionAnswerList(BaseModel):
-    pairs: list[QuestionAnswerAttributes]
-
-
-def generate_q_a_pairs(llm: str, message: str) -> QuestionAnswerList:
-    response = (
-        completion(
-            model=llm,
-            messages=[
-                {
-                    "content": GENERATE_QUESTION_ANSWER_PROMPT,
-                    "role": "system",
-                },
-                {
-                    "content": message,
-                    "role": "user",
-                },
-            ],
-            response_format=QuestionAnswerList,
-            temperature=app_config.temperature,
-            **completion_args(llm),
-        )
-        .choices[0]
-        .message.content
-    )
-
-    logger.info("Generated Question Answer: %s", response)
-
-    response_as_json = json.loads(response)
-    return QuestionAnswerList.model_validate(response_as_json)
 
 
 class MessageAttributes(BaseModel):
